Change datatype for the constructor's parameters
Save some memory, PIN-values won't be above 255, so a byte is more efficient than an int.

+++ SNESpaduino.cpp
@@ -12,7 +12,7 @@
 #include "SNESpaduino.h"
 
 // Constructor: Init pins
-SNESpaduino::SNESpaduino(int latch, int clock, int data)
+SNESpaduino::SNESpaduino(byte latch, byte clock, byte data)
 {
 	// Store the latch, clock and data pin for later use
 	PIN_LATCH = latch;
